requirements.txt
flask==3.0.2
flask-cors==4.0.1
groq==0.9.0
python-dotenv==1.0.1
requests==2.31.0
